refactor: reuse create page for i18n
